+++ webview/src/extension/extensionVariables.ts
