+++ webview/src/index.ts
@@ -3,8 +3,8 @@
 *  Licensed under the MIT License. See License.md in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-export * from './ConfirmationViewController';
-export * from './extensionVariables';
-export * from './LoadingViewController';
-export * from './OpenConfirmationViewStep';
-export * from './OpenLoadingViewStep';
+export * from './extension/ConfirmationViewController';
+export * from './extension/extensionVariables';
+export * from './extension/LoadingViewController';
+export * from './extension/OpenConfirmationViewStep';
+export * from './extension/OpenLoadingViewStep';
