+++ webview/src/webview/styles/confirmationView.scss
