+++ webview/src/webview/styles/loadingView.scss
