+++ webview/src/webview/theme/themeGenerator.ts
