+++ webview/src/webview/theme/utils/hueMap.ts
