+++ webview/src/webview/theme/utils/palettes.ts
